Broke a lot of the framework and restructued.
I've now made the framework follow MVC and will be fixing errors with
the framework that I've made through my refactors.
Future goals are to implement PSR-7 and working with a Router to handle
requests.
A man idea to keep in mind is also to make sure to export JSON as the
default response in future as it will allow a more 'dynamic' frontend to
be used.